Rename m_inference to membership_inf
